Discuz! X2 Release 20120329 +======================================+ ADD 增加應用中心(移植自X2.5) ADD 移植旋風存儲到X2 ADD 後台增加騰訊安全數據報表 ADD 增加騰訊安全的上報與處理接口 ADD 新增熱門話題、相關帖子功能
FIX 手機版查看專題頁面時的錯誤提示
FIX 用戶刪除了審核中的圖片，後台還是會提示
FIX 發表活動帖沒發圖片，提交後編輯添加活動圖片，無法添加
FIX 加載當前用戶組緩存時KEY判斷存在誤判的問題
FIX 防水牆：回帖白名單判斷問題

FIX 頻道文章原模塊為空時DIY操作保存失效
FIX 發送提醒支持HTML
FIX 暫時去除騰訊安全圖標的鏈接
FIX 修正邀請說明中HTML加斜桿的問題
FIX 添加插件升級文件，修改插件名稱

FIX 單主題(附件)最高收入設置無效
FIX 縱橫搜索去除多餘代碼
FIX UCenter修改用戶名問題 12014
FIX 搶樓帖到達規定樓層後主題已經關閉後刷新問題
FIX QQ互聯去掉群組裡面多餘的嵌入點

FIX 完善目錄可寫的檢測
FIX 遊客發的審核帖，遊客將不再能查看
FIX 站外url不解析時的BUG
FIX 手機版頂部導航問題
FIX 後台回收站翻頁bug

FIX 個人空間DIY模板問題
FIX 修正 IE8 下文章列表縮略圖 BUG
FIX 帖子推送到文章中的圖片不再進行權限限制，均可查看
FIX 後台日誌管理分頁參數
FIX DIY模塊  特殊主題過濾 的默認值

FIX 默認不做模板中是否包含圖片的檢查
FIX 優化onSiteGetAllUsers方法查詢
FIX 分類信息表單搜索的結果翻頁帶參數
FIX 模塊 summary長度無效
FIX threadimage tid 重複插入

FIX 因input的autofocus屬性導致在chrome瀏覽器中自動跳轉到頂部的問題
FIX QQ互聯feedlog 表同步到空間和微博的 status合併
FIX 優化QQ互聯微博簽名檔和解綁功能
FIX 管理面板刪帖不進回收站BUG
FIX 看帖頁站外分享重用論壇現有查詢，不再查數據庫

FIX 變量名seccode變量名錯誤
FIX static設置為http開頭的url時的bug
FIX 修復在帖子查看頁顯示的資料項不顯示的問題
FIX 短消息舉報
FIX 標題長度為空時，標題值為空

FIX 後台點擊訂單狀態無法查看 fix
FIX 郵箱統一判斷為32位
FIX 分類信息表單 XSS 威脅
FIX 用戶信息修改問題
FIX 家園同步表情

FIX 修正當設置閱讀權限時清容Feed主體內容
FIX 遠程調用計劃任務，不更新apc cache
FIX 群組帖子查看頁的DIY功能
FIX 用戶組到期後的處理
FIX 當版塊開啟圖片列表模式，在其版塊下也設置了分類信息【上傳圖片類型】時，編輯出錯

FIX 增加後台禁言時提醒用戶
FIX 審核回復不參與回帖獎勵
FIX 1221版本的一個斑竹刪帖的bug
FIX 個人空間分享頁面 刪除問題
FIX 修正啟用域名且同時啟用RW時帖子多入口的問題

FIX 在線時間顯示的bug
FIX IE9下inserttext修正
FIX 縱橫搜索拼接熱搜詞url時沒有對參數值進行rawurlencode的bug
FIX 論壇補充功能關閉了，還能補充
FIX 門戶添加頻道後，生成的頁面刪除欄目後無法刪除該頻道頁面。

FIX 計劃任務cron_cleanup_monthly.php多餘代碼
FIX 發送通知 郵件+html 不能解析html代碼 fix
FIX 插件模板解析問題
FIX 修改熱門話題展示樣式
FIX 修復 「朵」 「楔」等字結束添加\的問題

FIX 修正未啟用相冊分類時不顯示相冊分類
FIX 當沒有啟用相冊分類時的引發的BUG

+++ discuzx/upload/source/include/post/post_editpost.php
@@ -4,7 +4,7 @@
  *      [Discuz!] (C)2001-2099 Comsenz Inc.
  *      This is NOT a freeware, use is subject to license terms
  *
- *      $Id: post_editpost.php 25514 2011-11-14 02:37:37Z monkey $
+ *      $Id: post_editpost.php 29004 2012-03-22 04:04:58Z chenmengshu $
  */
 
 if(!defined('IN_DISCUZ')) {
@@ -616,9 +616,9 @@ if(!submitcheck('editsubmit')) {
 
 			DB::query("UPDATE ".DB::table('forum_thread')." SET typeid='$typeid', sortid='$sortid', subject='$subject', readperm='$readperm', price='$price' $closedadd $authoradd $polladd $replycreditadd".($_G['forum_auditstatuson'] && $audit == 1 ? ",displayorder='0', moderated='1'" : ",displayorder='$displayorder'").", status='$thread[status]' WHERE tid='$_G[tid]'", 'UNBUFFERED');
 
-			$_G['tid'] > 1 && DB::query("UPDATE ".DB::table('forum_thread')." SET subject='$subject' WHERE closed='$_G[tid]'", 'UNBUFFERED');
+			$thread['closed'] > 1 && DB::query("UPDATE ".DB::table('forum_thread')." SET subject='$subject' WHERE tid='$thread[closed]'", 'UNBUFFERED');
 
-			$tagstr = modthreadtag($_G['gp_tags'], $_G[tid]);
+			$tagstr = modthreadtag($_G['gp_tags'], $_G['tid']);
 
 		} else {
 
@@ -716,11 +716,13 @@ if(!submitcheck('editsubmit')) {
 
 		if($special == 4 && $isfirstpost && $_G['group']['allowpostactivity']) {
 			$activityaid = DB::result_first("SELECT aid FROM ".DB::table('forum_activity')." WHERE tid='$_G[tid]'");
-			if($activityaid != $_G['gp_activityaid']) {
+			if($activityaid && $activityaid != $_G['gp_activityaid']) {
 				$attach = DB::fetch_first("SELECT attachment, thumb, remote, aid FROM ".DB::table(getattachtablebytid($_G['tid']))." WHERE aid='$activityaid'");
 				DB::query("DELETE FROM ".DB::table('forum_attachment')." WHERE aid='$activityaid'");
 				DB::query("DELETE FROM ".DB::table(getattachtablebytid($_G['tid']))." WHERE aid='$activityaid'");
 				dunlink($attach);
+			}
+			if($_G['gp_activityaid']) {
 				$threadimageaid = $_G['gp_activityaid'];
 				convertunusedattach($_G['gp_activityaid'], $_G['tid'], $pid);
 				DB::query("UPDATE ".DB::table('forum_activity')." SET aid='$_G[gp_activityaid]' WHERE tid='$_G[tid]'");
